added searchProductByName controller and simplified products fields in controllers
