+++ app/backend/.gitignore
@@ -3,3 +3,5 @@ node_modules
 .env
 
 /src/generated/prisma
+
+/src/generated/prisma
