--- app/backend/products/products.controllers.js
@@ -1,64 +1,0 @@
-import prisma from "../src/config/db";
-
-// GET /products - Get all products
-export const getAllProducts = async (req, res) => {
-  try {
-    const products = await prisma.product.findMany();
-    res.json(products);
-  } catch (error) {
-    res.status(500).json({ error: 'Failed to fetch products' });
-  }
-};
-
-// GET /products/:id - Get one product
-export const getProductById = async (req, res) => {
-  const { id } = req.params;
-  try {
-    const product = await prisma.product.findUnique({ where: { id: Number(id) } });
-    if (!product) return res.status(404).json({ error: 'Product not found' });
-    res.json(product);
-  } catch (error) {
-    res.status(500).json({ error: 'Failed to fetch product' });
-  }
-};
-
-// POST /products - Create a new product
-export const createProduct = async (req, res) => {
-  const { item, amount, currency } = req.body;
-  try {
-    const newProduct = await prisma.product.create({
-      data: { item, amount: parseFloat(amount), currency },
-    });
-    res.status(201).json(newProduct);
-  } catch (error) {
-    res.status(400).json({ error: 'Failed to create product' });
-  }
-};
-
-// PUT /products/:id - Update an existing product
-export const updateProduct = async (req, res) => {
-  const { id } = req.params;
-  const { amount, currency } = req.body;
-  try {
-    const updatedProduct = await prisma.product.update({
-      where: { id: Number(id) },
-      data: { amount: parseFloat(amount), currency },
-    });
-    res.json(updatedProduct);
-  } catch (error) {
-    res.status(400).json({ error: 'Failed to update product' });
-  }
-};
-
-// DELETE /products/:id - Delete a product
-export const deleteProduct = async (req, res) => {
-  const { id } = req.params;
-  try {
-    await prisma.product.delete({ where: { id: Number(id) } });
-    res.json({ message: 'Product deleted successfully' });
-  } catch (error) {
-    res.status(400).json({ error: 'Failed to delete product' });
-  }
-};
-
-
