--- app/backend/products/products.routes.js
@@ -1,8 +1,0 @@
-import { Router } from 'express';
-import { getAllProducts } from './products.controllers.js';
-
-const productsRouter = Router()
-
-products.get('/products', getAllProducts);
-
-export default productsRouter
